requirements.txt
# Core metadata extraction
exifread>=3.0.0
piexif>=1.1.3
Pillow>=10.0.0
pdfminer.six>=20221105
python-docx>=1.0.0
olefile>=0.46
hachoir>=3.2.0
mutagen>=1.47.0
ffmpeg-python>=0.2.0
py7zr>=0.20.0
rarfile>=4.1
pefile>=2023.2.7
lief>=0.13.0
beautifulsoup4>=4.12.0

# Correlation and analysis
networkx>=3.1
pyvis>=0.3.2

# Visualization
folium>=0.14.0
plotly>=5.17.0

# GUI
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
python-magic>=0.4.27
python-dateutil>=2.8.2


